Refactored server express app for testing
